+++ truffle/CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
     * Added new `get.../set...` methods postfixed by `Static` for exclusively accessing static frame slots.
     * Added new `copy.../clear...` methods postfixed by `Static` for exclusively copying and clearing static frame slots.
     * Static frame slots are intended for situations where the type of a variable in a frame slots is known ahead of time and does not need any type checks (e.g. in statically typed languages).
+* GR-36557 Introduced `--engine.InliningUseSize` which changes the code size approximation during inlining to an approximation of the size of the graph rather than just the node count. This option is false by default.
 
 ## Version 22.1.0